+++ spi/src/main/java/org/jboss/pnc/spi/datastore/predicates/ArtifactPredicates.java
@@ -37,11 +37,11 @@ import org.jboss.pnc.model.TargetRepository_;
  */
 public class ArtifactPredicates {
 
-    public static Predicate<Artifact> withBuildRecordId(Integer buildRecordId) {
+    public static Predicate<Artifact> withBuildRecordId(Long buildRecordId) {
         return (root, query, cb) -> cb.equal(root.join(Artifact_.buildRecord).get(BuildRecord_.id), buildRecordId);
     }
 
-    public static Predicate<Artifact> withDependantBuildRecordId(Integer buildRecordId) {
+    public static Predicate<Artifact> withDependantBuildRecordId(Long buildRecordId) {
         return (root, query, cb) -> {
             Join<Artifact, BuildRecord> buildRecords = root.join(Artifact_.dependantBuildRecords);
             return cb.equal(buildRecords.get(BuildRecord_.id), buildRecordId);
