+++ spi/src/main/java/org/jboss/pnc/spi/datastore/predicates/BuildRecordPushResultPredicates.java
@@ -28,14 +28,14 @@ import org.jboss.pnc.spi.datastore.repositories.api.Predicate;
  */
 public class BuildRecordPushResultPredicates {
 
-    public static Predicate<BuildRecordPushResult> forBuildRecordOrderByIdDesc(Integer buildRecordId) {
+    public static Predicate<BuildRecordPushResult> forBuildRecordOrderByIdDesc(Long buildRecordId) {
         return (root, query, cb) -> {
             query.orderBy(cb.desc(root.get(BuildRecordPushResult_.id)));
             return cb.equal(root.get(BuildRecordPushResult_.buildRecord).get(BuildRecord_.id), buildRecordId);
         };
     }
 
-    public static Predicate<BuildRecordPushResult> successForBuildRecord(Integer buildRecordId) {
+    public static Predicate<BuildRecordPushResult> successForBuildRecord(Long buildRecordId) {
         return (root, query, cb) -> cb.and(
                 cb.equal(root.get(BuildRecordPushResult_.buildRecord).get(BuildRecord_.id), buildRecordId),
                 cb.equal(root.get(BuildRecordPushResult_.status), BuildPushStatus.SUCCESS));
